requirements.txt
python-dotenv
mysql
flask
pyjwt